+++ index.js
@@ -3,32 +3,11 @@
 const globby = require('globby');
 const eslint = require('eslint');
 const config = require('./config');
-
-function normalize(argv) {
-  const defaultOptions = {
-    ignore: []
-  };
-
-  const options = Object.assign(defaultOptions, argv);
-
-  options.ignore = options.ignore.concat([
-    '**/node_modules/**',
-    '**/bower_components/**',
-    'coverage/**',
-    '{tmp,temp}/**',
-    '**/*.min.js',
-    '**/bundle.js',
-    '{test,tests,spec,__tests__}/fixture{s,}/**',
-    'vendor/**',
-    'dist/**'
-  ]);
-
-  return options;
-}
+const option = require('./option');
 
 exports.lintText = (text, argv) => {
   const engine = new eslint.CLIEngine(config);
-  const options = normalize(argv);
+  const options = option.normalize(argv);
 
   return engine.executeOnText(text, options.fileName);
 };
@@ -36,7 +15,7 @@ exports.lintText = (text, argv) => {
 exports.lintFiles = (args, argv) => {
   const engine = new eslint.CLIEngine(config);
   const paths = args.length ? args : '**/*.js';
-  const options = normalize(argv);
+  const options = option.normalize(argv);
 
   return globby(paths, {
     ignore: options.ignore
